update to uio.h

+++ trunk/KERN/netinet/sctp_uio.h
@@ -704,10 +704,10 @@ struct sctp_mbuf_log {
 };
 
 struct sctp_cwnd_log {
-	uint32_t time_event;
-	uint8_t from;
-	uint8_t event_type;
-	uint8_t resv[2];
+	uint64_t time_event;
+	uint8_t  from;
+	uint8_t  event_type;
+	uint8_t  resv[2];
 	union {
 		struct sctp_log_closing close;
 		struct sctp_blk_args blk;
